+++ scripts/vagrant/run_tasks.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
-
+# shellcheck disable=SC2104
+#
+# for shellcheck before 4.4, use:
+#
+#   export SHELLCHECK_OPTS="-e SC2104"
+#
 scripts_dir=${SIMP_BUILDER_scripts_dir:-/vagrant/scripts}
 custom_scripts_dir=${SIMP_BUILDER_custom_scripts_dir:-$scripts_dir/custom}
 script_log_dir=${SIMP_BUILDER_log_dir:-$scripts_dir/../logs/session_$$}
@@ -14,7 +19,8 @@ stage_users=(vagrant)
 # ordered list of users that can provide pre-stage or post-stage custom scripts
 IFS=',' read -r -a custom_users <<< "${SIMP_BUILDER_users:-root,vagrant}"
 
-# Return the names of all *executable* file that match \d\d_.*
+# Return the names of all *executable* file that match \d\d_.* as a
+# null-delimited string
 #
 # Criteria:
 #
@@ -33,8 +39,11 @@ IFS=',' read -r -a custom_users <<< "${SIMP_BUILDER_users:-root,vagrant}"
 # $1 = directory to search for scripts ${_scripts_dir}/${user}/${stage}.d
 find_scripts_in()
 {
-    find $1/ -executable  -type f  -regextype posix-egrep \
-       -regex '^.*/[0-9][0-9]_[^/]*'  ! -regex '^.*/[^/]*\.disabled$'  -print0
+  [[ ${DEBUG} -gt 1 ]] && echo "    f--  find_scripts_in(): \$1='$1'" 1>&2
+  find "$1/" -executable  -type f  -regextype posix-egrep \
+     -regex '^.*/[0-9][0-9]_[^/]*'  ! -regex '^.*/[^/]*\.disabled$'  -print0 1>&2
+  find "$1/" -executable  -type f  -regextype posix-egrep \
+     -regex '^.*/[0-9][0-9]_[^/]*'  ! -regex '^.*/[^/]*\.disabled$'  -print0
 }
 
 # Convert a string into an acceptable environment variable name
@@ -57,37 +66,36 @@ skip_if_env_var_is_no()
 {
   local section=${2:-section}
   local env_var; { read env_var; } < <(sanitize_to_env_var_name "${1}")
+  [[ ${DEBUG} -gt 1 ]] && echo "    |--  skip_if_env_var_is_no():  ${section}: $env_var='${!env_var}' (proceeding)" 1>&2
   if [ "${!env_var}" == 'no' ]; then
     echo "    |!!  WARNING: SKIPPING ${section} because ${env_var}=no"
     continue
-  else
-    [[ ${DEBUG} -gt 1 ]] && echo "    |--  skip_if_env_var_is_no():  ${section}: $env_var='${!env_var}' (proceeding)" 1>&2
   fi
 }
-
 
 # $1 = full path to executable
 run_script()
 {
-  [ $# -lt 1 ] && { printf "ERROR: '$0':\n\nusage:\n\t$0 FILE\n\n" && exit 3; }
+  [ $# -lt 1 ] && { printf "ERROR: '%s':\n\nusage:\n\t%s FILE\n\n" "$0" && exit 3; }
 
-  local script_dir=$(basename $(dirname $1))
-  local script_parent_dir=$(basename $(dirname $(dirname $1)))
+  local script_dir=$(basename $(dirname "$1"))
+  local script_parent_dir=$(basename $(dirname $(dirname "$1")))
+  local sudo=
+  [[ "${script_dir}" =~ ^root ]] && sudo="sudo -E "
 
-  [[ "${script_dir}" =~ ^root ]] && sudo="sudo -E " || sudo=
   if [[ "${dry_run_mode}" == 'yes' ]]; then
     echo "==  SIMP_BUILDER: [dry-run] would have executed " \
-      "${sudo}${script_parent_dir}/${script_dir}/$(basename ${1})"
+      "${sudo}${script_parent_dir}/${script_dir}/$(basename "${1}")"
     continue
   fi
-  printf "\n==  SIMP_BUILDER: executing '$1':\n\n"
-
   local _log_dir="${script_log_dir}/${script_dir}"
-  mkdir -p ${_log_dir} || echo "WARNING: could not create log dir at ${_log_dir}"
+  mkdir -p "${_log_dir}" || echo "WARNING: could not create log dir at ${_log_dir}"
+  printf "\n==  SIMP_BUILDER: executing '%s':\n\n"  "${sudo} $1"
 
-  [[ ${DEBUG} -gt 0 ]] &&  echo "    |-   run_script():  ${sudo} $1 |& tee '${_log_dir}/$(basename $1).log'"
-  [[ ${DEBUG} -gt 1 ]] &&  echo "    |--  run_script():  _log_dir=${_log_dir}"
-  ${sudo} $1 |& tee "${_log_dir}/$(basename $1).log"
+  [[ ${DEBUG} -gt 0 ]] && echo "    |-   run_script():  ${sudo} $1 |& tee '${_log_dir}/$(basename "$1").log'"
+  [[ ${DEBUG} -gt 1 ]] && echo "    |--  run_script():  _log_dir=${_log_dir}"
+  ${sudo} "$1" | tee "${_log_dir}/$(basename "$1").log"
+  [[ ${DEBUG} -gt 1 ]] && echo "    |--  run_script():  finished executing $sudo '$1' _log_dir=${_log_dir}"
 }
 
 
@@ -101,7 +109,7 @@ run_stage()
   _scripts_dir=${3:-${custom_scripts_dir}}
   skip_if_env_var_is_no "SIMP_BUILDER__stage_${stage}" stage
 
-  [[ "${DEBUG}" -gt 0 ]] && echo "    |--  run_stage(): stage: ${stage}, users ("${#users[@]}"): ${users[@]}"
+  [[ "${DEBUG}" -gt 0 ]] && echo "    |--  run_stage(): stage: ${stage}, users (${#users[@]}): ${users[*]}"
 
   for user in "${users[@]}"; do
     skip_if_env_var_is_no "SIMP_BUILDER__stage_${stage}__user_${user}" stage
@@ -109,7 +117,6 @@ run_stage()
 
     export SIMP_BUILDER_user="${user}"
     local stage_user_scripts_dir="${_scripts_dir}/${user}/${stage}.d"
-    local n=0
 
     if [ -d "${stage_user_scripts_dir}" ]; then
       [[ "${DEBUG}" -gt 0 ]] && echo "  --+--  STAGE $stage   USER $user"
@@ -118,11 +125,14 @@ run_stage()
       continue
     fi
 
+    local n=0
     # Execute each script in order
     find_scripts_in "${stage_user_scripts_dir}" | sort -z | while read -d $'\0' file; do
+      [[ "${DEBUG}" -gt 0 ]] && echo "  --+--  run_stage(): STAGE $stage   USER $user  FILE: '${file}' (#${n})"
       file_name=$(basename "${file}")
       skip_if_env_var_is_no "SIMP_BUILDER__script_${file_name}" script
       skip_if_env_var_is_no "SIMP_BUILDER__stage_${stage}__user_${user}__script_${file_name}"
+      [[ "${DEBUG}" -gt 0 ]] && echo "    r--  run_stage(): STAGE $stage   USER $user:  \`run_script '${file}'\` (#${n})"
       run_script "${file}"
       (( ++n ))
     done
@@ -161,7 +171,7 @@ shift $((OPTIND -1))
 
 for task in "${tasks[@]}"; do
   printf "\n\n    ==+======================================\n"
-  printf     "               TASK: ${task}\n"
+  echo       "               TASK: ${task}"
   printf     "    ==+======================================\n\n"
   skip_if_env_var_is_no "SIMP_BUILDER__task_${task}" task