+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-pre2] - 2018-07-26
+Cleanups
+
+### Added
+- new `clone` stage added between `setup` and `build`
+- Workaround for simp-core#522 silliness
+  - (`rm rpm_docker/metadata.yml` in `vagrant/build.d/20_build_iso.sh`)
+
+### Changed
+- Scripts in `*.d/` dirs run in order
+- RVM install tries another gpg key servers if the first one fails
+- Documented customization options in README.md
+- Better shellcheck coverage in run_tasks.sh
+
 ## [1.0.0-pre1] - 2018-03-24
 Rewrote everything to use documented SIMP build process and template CI runners
 
@@ -19,7 +33,6 @@ Rewrote everything to use documented SIMP build process and template CI runners
 - Installs docker
 
 ### Changed
-
 - Complete re-write for SIMP 6.X
 
 ## [0.2.0] - 2016-03-03